codebase/pyproject.toml
[project]
requires-python = ">= 3.10.12"
name = "myGANsProject"
version = "1.1.1"
readme = "README.md"
license = { text = "myGANsProject © 2024 by RSLab@DISI, UniTrento (Italy) is licensed under CC BY 4.0 " }
description = "Python project with Docker Container dev environment and src layout"
authors = [{ name = "Samuele Trainotti", email = "samuele.trainotti@studenti.unitn.it" }]
maintainers = [{ name = "Name Surname", email = "your.email@domain.org" }]
dependencies = [
    "h5py == 3.10.0",
    "ipykernel == 6.29.4",
    "ipywidgets == 8.1.2",
    "PyYAML == 6.0.1",
    "rasterio == 1.3.9",
    "pdr == 1.4.0",
    "imageio == 2.37.0",
    "torchmetrics == 1.7.4",
    "seaborn == 0.13.2",
    "optuna == 4.4.0",
    "lpips == 0.1.4",
    "torch == 2.1.0",
    "torchvision == 0.16.0",
    "scikit-image == 0.22.0",
    "matplotlib == 3.8.2",
    "torch-fidelity == 0.3.0",
    "pandas",
    "scikit-learn",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
